# vendor-service/openapi.yaml
openapi: 3.0.3
info:
  title: Vendor Service API
  description: Vendor and storefront management
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local development
...
paths:
...
  /api/v1/vendors/{id}/documents:
    get:
      tags: [Documents]
      summary: Get vendor documents
      operationId: getVendorDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Documents list
  /api/v1/vendors/{id}/documents/{bucket}/{path}:
    delete:
      tags: [Documents]
      summary: Delete vendor document
      operationId: deleteVendorDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: bucket
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT